// src/app/models/user/UserMore.model.ts
export interface UserMore {
  uid: string;
  email: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  companyOrPrivateAddress: string;
  numberOfIdentityCard: string;
  companyRegistrationNumber: string;
  telephone: string;
  fax: string;
  homePage: string;
  assignedPostalCode: string;
  role: string[];
  roles: {};
  isActive: boolean;
  createdOn: Date;
}